add unit test and modify some interfaces
* WaterLevelTest
* ClosedStatusTest
* CycleControlTest
* DishWashingCyclePhase
* HeatingZoneTest
* LaundryCyclePhase
* OvenCyclePhase
* RapidModeTest
* RemoteControllabilityTest
* SoilLevelTest
* SpinSpeedLevelTest
* TimerTest

+++ cpp/samples/BaseSample/Controllee/AudioVolumeListener.cc
@@ -21,7 +21,7 @@ using namespace std;
 
 QStatus AudioVolumeListener::OnSetVolume(const uint8_t& volume)
 {
-    cout << "AudioVolumeListener::OnSetVolume() - volume : " << volume << endl;
+    cout << "AudioVolumeListener::OnSetVolume() - volume : " << (int)volume << endl;
     return ER_OK;
 }
